Move all helper functions to helper file
